move search-method to SearchController from HelloController.
