+++ src/main/java/com/example/demo/hello/Controller/HelloController.java
@@ -4,14 +4,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.example.demo.hello.HelloService;
 import com.example.demo.hello.domain.Employee;
-import com.example.demo.hello.domain.user.entity.EmployeeId;
-import com.example.demo.hello.domain.user.service.IdsearchRequest;
 
 /**
  * @author maetaku
@@ -45,17 +42,7 @@ public class HelloController {
 	//db.htmlに画面遷移
 	return "hello/db";
 	}
-	//Mapper検索
-	@PostMapping("/hello/db2")
-	public String postDbRequest1(@ModelAttribute IdsearchRequest idsearchRequest, Model model) {
-	//1件検索
-	 EmployeeId idSearch = service.search(idsearchRequest);
-	//検索結果をModelに登録
-	model.addAttribute("searchOne", idSearch);
-	
-	//db.htmlに画面遷移
-	return "hello/db2";
-	}
+
 //	 @RequestMapping(value = "/hello/db2", method = RequestMethod.POST)
 //	  public String search(@ModelAttribute IdsearchRequest idsearchRequest, Model model) {
 //		 IdsearchRequest sampleId = new IdsearchRequest();
