+++ src/main/java/com/example/demo/hello/Controller/SearchController.java
@@ -6,9 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 
 import com.example.demo.hello.HelloService;
 import com.example.demo.hello.domain.user.entity.EmployeeId;
+import com.example.demo.hello.domain.user.service.IdsearchRequest;
 
 @Controller
 public class SearchController {
@@ -22,11 +25,21 @@ public class SearchController {
 		List<EmployeeId>selectAll = service.selectAll();
 		m.addAttribute("selectAll", selectAll);
 		return "hello/employeeList";
-
+	}
 //	READ
+	//Mapper検索
+	@PostMapping("/hello/db2")
+	public String postDbRequest1(@ModelAttribute IdsearchRequest idsearchRequest, Model model) {
+		//1件検索
+		EmployeeId idSearch = service.search(idsearchRequest);
+		//検索結果をModelに登録
+		model.addAttribute("searchOne", idSearch);
 
+		//db.htmlに画面遷移
+		return "hello/db2";
+	}
 //	CREAT
 //	UODATE
 //	DEALETE
-	}
+	
 }
